requirements.txt
matplotlib==3.7.5
tensorflow==2.13.0
openpyxl
streamlit==1.35.0
streamlit-image-select==0.6.0
openpyxl==3.1.2
pandas==2.0.3
pillow==10.3.0
keras==2.13.1
numpy==1.24.3